Comple file download start with csv
Completed portion that determines the document path and downloads
document. After document is downloaded I am needing to save and reload
it, unsure as to why that is. sorted out my surveys from the list,
sorted the surveys by store. Need to format information so that it is
useful. Need to sort the list and assign based on grouping by oldest.