allow vixxorep to generate csv lines. add reps name to sites, generate csv
Added a tocsv method to the agent, which basically just go through the
sites and gets their csv outputs. Which basically go through the survey
lines and get their csv outputs. Also added the agents name to the site
when they are assigned it. Started main method and got up to having a
completly built csv with nowhere to go.